#67
Fixing button flow
