+++ src/index.ts
@@ -10,14 +10,13 @@ import {
   RESTPostAPIChatInputApplicationCommandsJSONBody,
 } from "discord.js";
 import { config } from "./config";
-import { getButtonHandler } from "./buttonHandler";
+import { getButtonHandler } from "./buttons/buttonHandler";
 
 import { deployCommands } from "./deploy-commands";
 
 import * as path from 'path';
 import * as fs from 'fs';
-import { handleCancelSwitch, handleEndSeries, handleGenerateAnotherCode,
-   handleGenerateAnotherConfirm, handleSwitchSidesConfirm, command as tournamentCommand } from "./commands/tournament";
+import { command as tournamentCommand } from "./commands/tournament";
 
 type ActionWrapper = {
   execute: (interaction: Interaction) => Promise<void>
@@ -84,8 +83,6 @@ client.once("ready", async () => {
 
   // We should grab events from API HERE :D
 });
-
-const channelId = process.env.CHANNEL_ID!;
 
 client.on(Events.InteractionCreate, async interaction => {
   if (interaction.isButton()) {
